% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfpkm.R
\name{zFPKMTransformDF}
\alias{zFPKMTransformDF}
\title{zFPKM Transformation}
\usage{
zFPKMTransformDF(fpkmDF, plotZFPKM = TRUE, PlotFileName = NULL,
  FacetTitles = FALSE, PlotXfloor = -20)
}
\arguments{
\item{fpkmDF}{FPKM (or TPM) dataframe}

\item{plotZFPKM}{Plot the scaled log_2(FPKM) and fitted Gaussian densities to screen [default = TRUE]}

\item{PlotFileName}{Plot the densities to specified file (.png) [default = NULL]}

\item{FacetTitles}{use to label each facet with the sample name [default = FALSE]}

\item{PlotXfloor}{Lower limit for X axis (log2FPKM units) [default = -20] set to NULL to disable}
}
\value{
zFPKM data frame
}
\description{
Perform the zFPKM transform on RNA-seq FPKM data. This algorithm is
based on the publication by Hart et al., 2013 (Pubmed ID 24215113). Reference recommends
using zFPKM > -3 to select expressed genes.  Validated with encode open/closed promoter chromatin structure epigenetic
data on six of the ENCODE cell lines.  Works well for gene level data using FPKM or TPM. Does not appear to calibrate well
for transcript level data.
}
\examples{
zFPKMDat <- zFPKMTransformDF(MyFPKMdf)
zFPKMDat <- zFPKMTransformDF(MyFPKMdf, plotZFPKM=FALSE, PlotFileName="Figs/zFPKM.png")

}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/24215113}
}
\author{
Ron Ammar, \email{ron.ammar@bms.com}
}
\keyword{zFPKM}
