% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zfpkm.R
\name{zFPKMTransformDF}
\alias{zFPKMTransformDF}
\title{zFPKM Transformation}
\usage{
zFPKMTransformDF(fpkmDF, PlotFileName = "zFPKM.PNG", FacetTitles = FALSE,
  PlotDir = ".")
}
\arguments{
\item{fpkmDF}{FPKM (or TPM) dataframe}

\item{PlotFileName}{Provide a name for the plot file (.png) [default = zFPKM.PNG]}

\item{FacetTitles}{use to label each facet with the sample name [default = FALSE]}

\item{PlotDir}{Specify a subdirectory to place the plot in [default = "."]}
}
\value{
zFPKM data frame
}
\description{
Perform the zFPKM transform on RNA-seq FPKM data. This algorithm is
based on the publication by Hart et al., 2013 (Pubmed ID 24215113). Reference recommends
using zFPKM >-3 to select expressed genes.  Validated with encode open/closed promoter chromatin structure epigenetic
data on six of the ENCODE cell lines.  Works well for gene level data using FPKM or TPM.  Does not appear to calibrate well
for transcript level data.
}
\examples{
MyzFPKM_dataframe = zFPKMTransformDF (MyFPKMdf, PlotFileName = "MyZFPKM.png",
       FacetTitles = TRUE, PlotDir = "./MyOutputDir")
}
\author{
Ron Ammar, \email{ron.ammar@bms.com}
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/24215113}
}
\keyword{zFPKM}

